requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
langgraph
langchain
langchain-core
matplotlib
numpy>=2.0.0
pydantic
python-multipart
chromadb
langchain-community
streamlit
requests
pandas
tavily-python
langchain-groq
langgraph-checkpoint-postgres
sqlalchemy
psycopg[binary]
psycopg-pool
deepgram-sdk
websockets
python-dotenv
groq
langsmith
deepagents
langchain-google-genai
seaborn
duckduckgo-search
pillow
huggingface_hub
google-adk
langchain-mcp-adapters
mcp-tavily
instructor
langchain-huggingface
langsmith[otel]
